NoChangesReviewDbWrapper: Add strict mode and fail mutation ops
Most plugins in the wild do not support NoteDb yet. With current
implementation of NoChangesReviewDbWrapper, where all mutation
operations are just dummy methods without failing the caller, it
is hard to track down and understand the problem, because only
the subsequent errors show up.

This change unconditionally fails all database mutation operations.
The caller sites are adjusted correspondingly to skip the database
mutation operations when underlying database instance does not
support change tables mutations operations. Note, that it wouldn't
be sufficient to check for the migration state WRITE, because the
BatchUpdateReviewDb instance exposes mixed NoteDb/ReviewDb modes.

Bug: Issue 9681

+++ gerrit-server/src/main/java/com/google/gerrit/server/change/ConsistencyChecker.java
@@ -669,10 +669,12 @@ public class ConsistencyChecker {
       // Delete dangling key references.
       ReviewDb db = DeleteChangeOp.unwrap(ctx.getDb());
       accountPatchReviewStore.get().clearReviewed(psId);
-      db.changeMessages().delete(db.changeMessages().byChange(psId.getParentKey()));
-      db.patchSetApprovals().delete(db.patchSetApprovals().byPatchSet(psId));
-      db.patchComments().delete(db.patchComments().byPatchSet(psId));
-      db.patchSets().deleteKeys(Collections.singleton(psId));
+      if (db.changesTablesEnabled()) {
+        db.changeMessages().delete(db.changeMessages().byChange(psId.getParentKey()));
+        db.patchSetApprovals().delete(db.patchSetApprovals().byPatchSet(psId));
+        db.patchComments().delete(db.patchComments().byPatchSet(psId));
+        db.patchSets().deleteKeys(Collections.singleton(psId));
+      }
 
       // NoteDb requires no additional fiddling; setting the state to deleted is
       // sufficient to filter everything else out.
